:name_badge: Rename `StateObject`
Overview
===
Rename `StateObject` to `UseCase.ObservableObject`
And set type aliase ObservableUseCase
