+++ Sources/UseCaseKit/UseCase+Combine.swift
@@ -1,4 +1,4 @@
-// 
+//
 // Copyright © 2021 @crexista. All rights reserved.
 //
