+++ Tests/UseCaseKitTests/ObservableUseCaseSpec.swift
@@ -7,11 +7,11 @@ import Quick
 import Nimble
 
 @available(iOS 13.0, *)
-class StateObjectSpec: QuickSpec {
+class ObservableUseCaseSpec: QuickSpec {
 
     override func spec() {
         var usecase: UseCase<MockCommand>!
-        var stateObject: StateObject<MockCommand>!
+        var observableMock: ObservableUseCase<MockCommand>!
 
         describe("StateObject Subscribing Test") {
 
@@ -19,18 +19,18 @@ class StateObjectSpec: QuickSpec {
 
                 beforeEach {
                     usecase = .test(nil)
-                    stateObject = usecase.asStateObject()
+                    observableMock = usecase.asObservableObject()
                 }
 
                 it("StateObject's state is `initial`") {
-                    expect(stateObject.state) == .initial
+                    expect(observableMock.state) == .initial
                 }
 
                 it("`StateObject`'s state will change to mock1 after test1 command is dispatched to `UseCase`") {
                     waitUntil { end in
                         usecase.dispatch(.test1)
                         _ = usecase.state
-                        DispatchQueue.main.async { expect(stateObject.state) == .mock1; end() }
+                        DispatchQueue.main.async { expect(observableMock.state) == .mock1; end() }
                     }
                 }
 
@@ -38,7 +38,7 @@ class StateObjectSpec: QuickSpec {
                     waitUntil { end in
                         usecase.dispatch(.test2)
                         _ = usecase.state
-                        DispatchQueue.main.async { expect(stateObject.state) == .mock2; end() }
+                        DispatchQueue.main.async { expect(observableMock.state) == .mock2; end() }
                     }
                 }
 
@@ -51,20 +51,20 @@ class StateObjectSpec: QuickSpec {
 
                 beforeEach {
                     usecase = .test(nil)
-                    stateObject = usecase.asStateObject()
+                    observableMock = usecase.asObservableObject()
                 }
 
                 it("`UseCase` state will change to mock1 after test1 command is dispatched to `StateObject`") {
                     waitUntil { end in
                         usecase.sink(ignoreFirst: true) { expect($0) == .mock1; end() }
-                        stateObject.dispatch(.test1)
+                        observableMock.dispatch(.test1)
                     }
                 }
 
                 it("`UseCase` state will change to mock2 after test2 command is dispatched to `StateObject`") {
                     waitUntil { end in
                         usecase.sink(ignoreFirst: true) { expect($0) == .mock2; end() }
-                        stateObject.dispatch(.test2)
+                        observableMock.dispatch(.test2)
                     }
                 }
